+++ piecrust/tests/box.rs
@@ -8,6 +8,7 @@ use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
 use rkyv::{check_archived_root, Deserialize, Infallible};
 
 const OWNER: [u8; 32] = [0u8; 32];
+const LIMIT: u64 = 1_000_000;
 
 #[test]
 pub fn box_set_get() -> Result<(), Error> {
@@ -15,16 +16,19 @@ pub fn box_set_get() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let id = session
-        .deploy(contract_bytecode!("box"), ContractData::builder(OWNER))?;
+    let id = session.deploy(
+        contract_bytecode!("box"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
-    let value: Option<i16> = session.call(id, "get", &())?.data;
+    let value: Option<i16> = session.call(id, "get", &(), LIMIT)?.data;
 
     assert_eq!(value, None);
 
-    session.call::<i16, ()>(id, "set", &0x11)?;
+    session.call::<i16, ()>(id, "set", &0x11, LIMIT)?;
 
-    let value = session.call::<_, Option<i16>>(id, "get", &())?.data;
+    let value = session.call::<_, Option<i16>>(id, "get", &(), LIMIT)?.data;
 
     assert_eq!(value, Some(0x11));
 
@@ -37,18 +41,21 @@ pub fn box_set_get_raw() -> Result<(), Error> {
 
     let mut session = vm.session(SessionData::builder())?;
 
-    let id = session
-        .deploy(contract_bytecode!("box"), ContractData::builder(OWNER))?;
+    let id = session.deploy(
+        contract_bytecode!("box"),
+        ContractData::builder(OWNER),
+        LIMIT,
+    )?;
 
-    let value_bytes = session.call_raw(id, "get", vec![])?.data;
+    let value_bytes = session.call_raw(id, "get", vec![], LIMIT)?.data;
     let value = deserialize_value(&value_bytes)?;
 
     assert_eq!(value, None);
 
     let value_bytes = serialize_value(0x11)?;
-    session.call_raw(id, "set", value_bytes)?;
+    session.call_raw(id, "set", value_bytes, LIMIT)?;
 
-    let value_bytes = session.call_raw(id, "get", vec![])?.data;
+    let value_bytes = session.call_raw(id, "get", vec![], LIMIT)?.data;
     let value = deserialize_value(&value_bytes)?;
 
     assert_eq!(value, Some(0x11));